Move scan builders to core/utils/accumulo-utils
