+++ core/utils/accumulo-utils/src/main/java/datawave/scan/ExecutionHintHelper.java
@@ -1,4 +1,4 @@
-package datawave.query.scan;
+package datawave.scan;
 
 import java.util.Map;
