+++ core/utils/accumulo-utils/src/test/java/datawave/scan/ExecutionHintHelperTest.java
@@ -1,4 +1,4 @@
-package datawave.query.scan;
+package datawave.scan;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
