+++ warehouse/core/pom.xml
@@ -44,11 +44,6 @@
             <artifactId>commons-cli</artifactId>
         </dependency>
         <dependency>
-            <!-- Accumulo BatchScanner requires IteratorChain class -->
-            <groupId>commons-collections</groupId>
-            <artifactId>commons-collections</artifactId>
-        </dependency>
-        <dependency>
             <groupId>gov.nsa.datawave</groupId>
             <artifactId>datawave-common</artifactId>
             <version>${project.version}</version>
