+++ warehouse/query-core/src/main/java/datawave/query/jexl/lookups/FieldedRegexIndexLookup.java
@@ -20,9 +20,9 @@ import com.google.common.base.Preconditions;
 
 import datawave.core.iterators.FieldedRegexExpansionIterator;
 import datawave.query.config.ShardQueryConfiguration;
-import datawave.query.scan.ExecutionHintHelper;
-import datawave.query.scan.ScannerBuilder;
 import datawave.query.tables.ScannerFactory;
+import datawave.scan.ExecutionHintHelper;
+import datawave.scan.ScannerBuilder;
 import datawave.util.time.DateHelper;
 
 /**
