+++ warehouse/query-core/src/test/java/datawave/query/tables/ScanSessionManagerTest.java
@@ -34,9 +34,9 @@ import datawave.core.query.configuration.GenericQueryConfiguration;
 import datawave.core.query.configuration.QueryData;
 import datawave.microservice.query.Query;
 import datawave.microservice.query.QueryImpl;
-import datawave.query.scan.BatchScannerBuilder;
-import datawave.query.scan.ScannerBuilder;
 import datawave.query.tables.async.ScannerChunk;
+import datawave.scan.BatchScannerBuilder;
+import datawave.scan.ScannerBuilder;
 import datawave.util.TableName;
 
 public class ScanSessionManagerTest {
